% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cap_names}
\alias{cap_names}
\title{Capitalize set descriptions}
\usage{
cap_names(x)
}
\arguments{
\item{x}{A character vector of term descriptions (i.e. the
\code{gs_description} column of \code{\link{msigdbr}} results).}
}
\value{
A character vector of conditionally-capitalized terms.
}
\description{
Capitalizes the first letter of a description, unless the first
word was not entirely lowercase. For example, the first letter of
"magnesium ion binding" is capitalized, but "tRNA binding" would remain
unchanged. This should preserve special capitalization while slightly
improving the appearance of descriptions in visualizations.
}
\examples{
\donttest{
x <- msigdbr(species = "Homo sapiens",
             category = "C5",
             subcategory = "GO:MF")
x <- update_GO_names(x)
head(unique(x$gs_description)) # before


x$gs_description <- cap_names(x$gs_description)
head(unique(x$gs_description)) # after
}
}
\author{
Tyler Sagendorf
}
