% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-msigdb-df.R
\name{load_msigdb_df}
\alias{load_msigdb_df}
\title{Download, cache, and load the gene sets database}
\usage{
load_msigdb_df(
  target_species = c("HS", "MM"),
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{target_species}{Species abbreviation for human or mouse databases (\code{"HS"} or \code{"MM"}).}

\item{overwrite}{A logical indicating whether existing cached file should be overwritten.}

\item{verbose}{A logical indicating whether to print progress information.}
}
\value{
A data frame of gene sets and their member genes.
}
\description{
Download, cache, and load the gene sets database
}
