% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbr-db.R
\name{msigdbr_db}
\alias{msigdbr_db}
\title{Load the gene sets database}
\usage{
msigdbr_db(target_species = c("HS", "MM"), overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{target_species}{Species abbreviation for human or mouse databases (\code{"HS"} or \code{"MM"}).}

\item{overwrite}{A logical indicating whether existing cached file should be overwritten.}

\item{verbose}{A logical indicating whether to print progress information.}
}
\value{
A data frame of gene sets and their member genes.
}
\description{
Download, cache, and load the gene sets data frame.
}
