% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{msigdbr_show_species}
\alias{msigdbr_show_species}
\title{List the species available in the msigdbr package}
\usage{
msigdbr_show_species()
}
\value{
A vector of possible species.
}
\description{
This function is being deprecated and replaced by \code{msigdbr_species()}.
}
