% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbr-species.R
\name{msigdbr_species}
\alias{msigdbr_species}
\title{List the species available in the msigdbr package}
\usage{
msigdbr_species()
}
\value{
A data frame of the available species.
}
\description{
List the species available in the msigdbr package
}
\examples{
msigdbr_species()
}
