% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_database.R
\name{update_database}
\alias{update_database}
\title{Update database to be the latest released in \href{https://data.broadinstitute.org/gsea-msigdb/msigdb/}{msigdb website}}
\usage{
update_database(
  mdb_version0 = "2023.1",
  species_df = data.frame(species_vereison = c("Hs", "Mm"), species_annotation =
    c("Human", "Mouse"))
)
}
\arguments{
\item{mdb_version0}{a string which is version name shown in \href{https://data.broadinstitute.org/gsea-msigdb/msigdb/release/}{msigdb releases},
but the characters specifying the species are removed}

\item{species_df}{a data.frame with two columns, the 1st column is the
character specifying species in full version name,
the 2nd column is the character specifying species in the name of
annotation files.}
}
\value{
No return, but will save rda files in /data.
}
\description{
Update database to be the latest released in \href{https://data.broadinstitute.org/gsea-msigdb/msigdb/}{msigdb website}
}
\examples{
# update database to be version "2023.1", and get both human and mouse from \href{https://data.broadinstitute.org/gsea-msigdb/msigdb/}{msigdb website}
\donttest{
update_database(mdb_version0="2023.1",
                species_df=data.frame(
                      'species_vereison' = c('Hs','Mm'),
                      'species_annotation' = c('Human','Mouse')))
}

}
