% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_GO_names}
\alias{update_GO_names}
\title{Update Gene Ontology term descriptions}
\usage{
update_GO_names(x, version = packageVersion("msigdbr"))
}
\arguments{
\item{x}{\code{data.frame} produced by \code{\link{msigdbr}} containing columns
\code{gs_description} and \code{gs_subcat}.}

\item{version}{Character string specifying the version of \code{msigdbr} to use. Defaults to the current version.}
}
\value{
A \code{data.frame}. The same as \code{x}, but with updated descriptions.
}
\description{
Update entries in the \code{gs_description} column of
\code{\link{msigdbr}} results to use terms from the
\href{http://geneontology.org/}{Gene Ontology Consortium}.
}
\details{
This function assumes that the phrase "GO-basic obo file released
on" is present in the MSigDB release notes for that version and is followed
by a date.
}
\examples{
\donttest{
x <- msigdbr(species = "Homo sapiens",
             category = "C5",
             subcategory = "GO:MF")
head(unique(x$gs_description)) # before

x <- update_GO_names(x)
head(unique(x$gs_description)) # after
}
}
\references{
Ashburner, M., et al. (2000). Gene ontology: tool for the
unification of biology. The Gene Ontology Consortium. \emph{Nature genetics},
\emph{25}(1), 25–29. \url{https://doi.org/10.1038/75556}

Gene Ontology Consortium (2021). The Gene Ontology resource: enriching a
GOld mine. \emph{Nucleic acids research}, \emph{49}(D1), D325–D334.
\url{https://doi.org/10.1093/nar/gkaa1113}
}
\seealso{
\code{\link{cap_names}}, \href{https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Release_Notes}{MSigDB
Release Notes}, \href{http://release.geneontology.org/}{Gene Ontology Data
Archive}
}
\author{
Tyler Sagendorf
}
