% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbr-package.R
\docType{package}
\name{msigdbr-package}
\alias{msigdbr-package}
\title{msigdbr: MSigDB Gene Sets for Multiple Organisms in a Tidy Data Format}
\description{
Provides the 'Molecular Signatures Database' (MSigDB) gene sets typically used with the 'Gene Set Enrichment Analysis' (GSEA) software (Subramanian et al. 2005 \doi{10.1073/pnas.0506580102}, Liberzon et al. 2015 \doi{10.1016/j.cels.2015.12.004}, Castanza et al. 2023 \doi{10.1038/s41592-023-02014-7}) as an R data frame. The package includes the human genes as listed in MSigDB as well as the corresponding symbols and IDs for frequently studied model organisms such as mouse, rat, pig, fly, and yeast.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://igordot.github.io/msigdbr/}
  \item Report bugs at \url{https://github.com/igordot/msigdbr/issues}
}

}
\author{
\strong{Maintainer}: Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})

Authors:
\itemize{
  \item Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})
}

}
\keyword{internal}
