% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbr-check-data.R
\name{msigdbr_check_data}
\alias{msigdbr_check_data}
\title{Check that the data package is installed}
\usage{
msigdbr_check_data()
}
\description{
Check that the 'msigdbdf' data package is installed.
If not, provide instructions for installation.
A dependency listed in DESCRIPTION Suggests is not guaranteed to be installed.
}
