% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{msigdbr_collections}
\alias{msigdbr_collections}
\title{List the collections available in the msigdbr package}
\usage{
msigdbr_collections()
}
\value{
A data frame of the available collections.
}
\description{
List the collections available in the msigdbr package
}
\examples{
msigdbr_collections()
}
