% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbr-collections.R
\name{msigdbr_collections}
\alias{msigdbr_collections}
\title{List the collections available in the msigdbr package}
\usage{
msigdbr_collections(db_species = "Hs")
}
\arguments{
\item{db_species}{Species abbreviation for the human or mouse databases (\code{"Hs"} or \code{"Mm"}).}
}
\value{
A data frame of the available collections.
}
\description{
List the collections available in the msigdbr package
}
