% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{msigdbr_show_species}
\alias{msigdbr_show_species}
\title{List the species available in msigdbr.}
\usage{
msigdbr_show_species()
}
\value{
a vector of possible species
}
\description{
List the species available in msigdbr.
}
